Added comments. Refactored basic scheduler class.

+++ src/main/java/config/CanvasConstants.java
@@ -1,8 +1,9 @@
 package config;
 
+/** HTML Canvas related constants that is used on the client side. */
 public class CanvasConstants {
 	public static final int CANVAS_WIDTH = 800;
 	public static final int CANVAS_HEIGHT = 600;
-	public static final int CANVAS_HALF_WIDTH = CANVAS_WIDTH / 2;
-	public static final int CANVAS_HALF_HEIGHT = CANVAS_HEIGHT / 2;
+	public static final long CANVAS_HALF_WIDTH = CANVAS_WIDTH / 2;
+	public static final long CANVAS_HALF_HEIGHT = CANVAS_HEIGHT / 2;
 }
